### 2024-10-24 71a139c
minor fixes and updates to paper as suggested by @jromanowska
