+++ paper/paper.bib
@@ -54,6 +54,8 @@
   year={2004}
 }
 
+@article{Panch2021, title={The systems biology simulation core library}, volume={38}, ISSN={1367-4811}, url={http://dx.doi.org/10.1093/bioinformatics/btab669}, DOI={10.1093/bioinformatics/btab669}, number={3}, journal={Bioinformatics}, publisher={Oxford University Press (OUP)}, author={Panchiwala, Hemil and Shah, Shalin and Planatscher, Hannes and Zakharchuk, Mykola and König, Matthias and Dräger, Andreas}, editor={Xu, Jinbo}, year={2021}, month=sep, pages={864–865} }
+
 @book{Python,
  author = {Van Rossum, Guido and Drake, Fred L.},
  title = {Python 3 Reference Manual},
@@ -114,6 +116,8 @@ journal = {Journal of Statistical Mechanics: Theory and Experiment},
 }
 
 @article{Toni2008, title={Approximate Bayesian computation scheme for parameter inference and model selection in dynamical systems}, volume={6}, ISSN={1742-5662}, url={http://dx.doi.org/10.1098/rsif.2008.0172}, DOI={10.1098/rsif.2008.0172}, number={31}, journal={Journal of The Royal Society Interface}, publisher={The Royal Society}, author={Toni, Tina and Welch, David and Strelkowa, Natalja and Ipsen, Andreas and Stumpf, Michael P.H}, year={2008}, month=jul, pages={187–202} }
+
+@article{Welsh2022, title={libRoadRunner 2.0: a high performance SBML simulation and analysis library}, volume={39}, ISSN={1367-4811}, url={http://dx.doi.org/10.1093/bioinformatics/btac770}, DOI={10.1093/bioinformatics/btac770}, number={1}, journal={Bioinformatics}, publisher={Oxford University Press (OUP)}, author={Welsh, Ciaran and Xu, Jin and Smith, Lucian and König, Matthias and Choi, Kiri and Sauro, Herbert M}, editor={Martelli, Pier Luigi}, year={2022}, month=dec }
 
 @article{Wilkinson2009, title={Stochastic modelling for quantitative description of heterogeneous biological systems}, volume=10, ISSN={1471-0064}, url={http://dx.doi.org/10.1038/nrg2509}, DOI={10.1038/nrg2509}, number=2, journal={Nature Reviews Genetics}, publisher={Springer Science and Business Media LLC}, author={Wilkinson, Darren J.}, year=2009, month=feb, pages={122–133} }
 
