### 2024-10-24 #7329 @darrenjw
Thanks. The build is fixed and I've added references to a couple of existing systems biology simulation libraries.
