### 2024-10-25 #7329 @jromanowska
:wave: @cthoyt, @benlansdell, @ortega2247 & @antoineruzette, would any of you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
