### 2024-10-25 darrenjw/jax-smfsb#7 @cthoyt
Apply code formatter
The code in this repo doesn't conform to a consistent code style. I suggest applying ruff or black to standardize the code style in a way that is more typical for Python programs
