### 2024-10-25 darrenjw/jax-smfsb#8 @cthoyt
Use pythonic names
The repository uses very java-flavored names, which makes it less approachable for python programmers. You can read PEP-8's naming conventions guide to get familiar with https://peps.python.org/pep-0008/#naming-conventions

for example, `step = lvmod.stepGillespie()` feels very java. Python conventions would be to write it as `step = lvmod.step_gillespie()`

similarly, using the number 2 instead of `to` is also not very pythonic

https://github.com/darrenjw/jax-smfsb/blob/71a139c148c4b6c777c80a0bb71236a1fa834b6c/src/jsmfsb/smfsbSbml.py#L12

should be `mod_to_spn`, or ideally, something that is not an abbreviation

I understand that this might be because you're trying to mimic the interface of `smfsb`, but this is irrelevant in terms of reviewing this repo
