### 2024-10-25 darrenjw/jax-smfsb#9 @cthoyt
Add more documentation
Right now, the documentation seems to just be demo files in https://github.com/darrenjw/jax-smfsb/tree/main/demos. This isn't sufficient to motivate why the package exists and to onboard a new user - the documentation in https://jax-smfsb.readthedocs.io/en/latest/index.html should be more extensive and ideally linked with a ReadTheDocs badge like most python packages do.

Please watch Daniele Procida's great talk on how to think about writing good documentation: https://youtu.be/azf6yzuJt54?si=GzIWdr8-24KCXfVP
