### 2024-10-25 #7329 @cthoyt
I'd made some quick suggestions to @darrenjw to try and update the package to conform to Python community code standards (https://github.com/darrenjw/jax-smfsb/issues/7, https://github.com/darrenjw/jax-smfsb/issues/8) and to add some documentation for new users (https://github.com/darrenjw/jax-smfsb/issues/9)

I will consider reviewing the manuscript if these are addressed
