### 2024-10-25 #7329 @jromanowska
> I'd made some quick suggestions to @darrenjw to try and update the package to conform to Python community code standards ([darrenjw/jax-smfsb#7](https://github.com/darrenjw/jax-smfsb/issues/7), [darrenjw/jax-smfsb#8](https://github.com/darrenjw/jax-smfsb/issues/8)) and to add some documentation for new users ([darrenjw/jax-smfsb#9](https://github.com/darrenjw/jax-smfsb/issues/9))
>
> I will consider reviewing the code further and the manuscript if these are addressed

Thank you so much for looking into that so fast! We'll wait for the author to address these issues. :+1:
