### 2024-10-30 #7329 @jromanowska
Thank you, @antoineruzette. We'll wait some days for the response from @darrenjw  to [issues raised](https://github.com/openjournals/joss-reviews/issues/7329#issuecomment-2437204376) by @cthoyt.

In the meantime, if you or @benlansdell have any other comments, please write them here. Note that this is not yet the official review thread, so you don't need to look into the code yet.
