### 2024-11-03 3835819
starting on syntax refactor
