+++ tests/test_spn.py
@@ -7,7 +7,7 @@ import jax.numpy as jnp
 
 def test_spn():
     lv = jsmfsb.models.lv()
-    step = lv.stepGillespie()
+    step = lv.step_gillespie()
     k0 = jax.random.key(42)
     x1 = step(k0, lv.m, 0, 1)
     assert(x1.shape == (2,))
@@ -15,7 +15,7 @@ def test_spn():
 
 def test_pts():
     lv = jsmfsb.models.lv()
-    step = lv.stepPTS(0.001)
+    step = lv.step_poisson(0.001)
     k0 = jax.random.key(42)
     x1 = step(k0, lv.m, 0, 1)
     assert(x1.shape == (2,))
@@ -23,7 +23,7 @@ def test_pts():
 
 def test_euler():
     lv = jsmfsb.models.lv()
-    step = lv.stepEuler(0.001)
+    step = lv.step_euler(0.001)
     k0 = jax.random.key(42)
     x1 = step(k0, lv.m, 0, 1)
     assert(x1.shape == (2,))
@@ -31,7 +31,7 @@ def test_euler():
 
 def test_cle():
     lv = jsmfsb.models.lv()
-    step = lv.stepCLE(0.001)
+    step = lv.step_cle(0.001)
     k0 = jax.random.key(42)
     x1 = step(k0, lv.m, 0, 1)
     assert(x1.shape == (2,))
