### 2024-11-03 9e1ea0e
rename methods to be more pythonic, in response to issue #8
