+++ demos/stepGillespie2D.py
@@ -12,7 +12,7 @@ T=10
 x0 = jnp.zeros((2,M,N))
 lv = jsmfsb.models.lv()
 x0 = x0.at[:,int(M/2),int(N/2)].set(lv.m)
-stepLv2d = lv.step_gillespie2D(jnp.array([0.6, 0.6]))
+stepLv2d = lv.step_gillespie_2d(jnp.array([0.6, 0.6]))
 k0 = jax.random.key(42)
 x1 = stepLv2d(k0, x0, 0, T)
 
@@ -20,7 +20,7 @@ fig, axis = plt.subplots()
 for i in range(2):
     axis.imshow(x1[i,:,:])
     axis.set_title(lv.n[i])
-    fig.savefig(f"step_gillespie2D{i}.pdf")
+    fig.savefig(f"step_gillespie_2d{i}.pdf")
 
 
 # eof
