+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "hatchling.build"
 
 [project]
 name = "jsmfsb"
-version = "1.1.0"
+version = "1.1.1"
 authors = [
   { name="Darren Wilkinson", email="darrenjwilkinson@btinternet.com" },
 ]
