### 2024-11-03 darrenjw/jax-smfsb#8 @darrenjw
OK, I've had a go at this. I don't like unnecessary breaking changes, but I do get your point. I've tried to make consistent changes to both this library and `smfsb`. Of course they are now both somewhat inconsistent with the R library and the Scala library, but I guess that's inevitable.
