+++ demos/m-mm.py
@@ -18,12 +18,12 @@ print(step(k0, mmmod.m, 0, 30))
 out = jsmfsb.sim_time_series(k0, mmmod.m, 0, 100, 0.1, step)
 
 import matplotlib.pyplot as plt
+
 fig, axis = plt.subplots()
 for i in range(4):
-	axis.plot(range(out.shape[0]), out[:,i])
+    axis.plot(range(out.shape[0]), out[:, i])
 
 axis.legend(mmmod.n)
 fig.savefig("m-mm.pdf")
 
 # eof
-
