+++ demos/time-lv-cle.py
@@ -18,7 +18,7 @@ k0 = jax.random.key(42)
 ## Start timer
 startTime = time.time()
 out = jsmfsb.sim_sample(k0, 10000, lvmod.m, 0, 20, step)
-#out = jsmfsb.sim_sampleMap(k0, 10000, lvmod.m, 0, 20, step)
+# out = jsmfsb.sim_sampleMap(k0, 10000, lvmod.m, 0, 20, step)
 endTime = time.time()
 ## End timer
 elapsedTime = endTime - startTime
@@ -26,12 +26,12 @@ print(f"\n\nElapsed time: {elapsedTime} seconds\n\n")
 
 out = jnp.where(out > 1000, 1000, out)
 import scipy as sp
+
 print(sp.stats.describe(out))
-fig, axes = plt.subplots(2,1)
+fig, axes = plt.subplots(2, 1)
 for i in range(2):
-    axes[i].hist(out[:,i], bins=50)
+    axes[i].hist(out[:, i], bins=50)
 fig.savefig("time-lv-cle.pdf")
 
 
 # eof
-
