+++ demos/time-lv-gillespie.py
@@ -25,12 +25,12 @@ print(f"\n\nElapsed time: {elapsedTime} seconds\n\n")
 
 out = jnp.where(out > 1000, 1000, out)
 import scipy as sp
+
 print(sp.stats.describe(out))
-fig, axes = plt.subplots(2,1)
+fig, axes = plt.subplots(2, 1)
 for i in range(2):
-    axes[i].hist(out[:,i], bins=50)
+    axes[i].hist(out[:, i], bins=50)
 fig.savefig("time-lv-gillespie.pdf")
 
 
 # eof
-

[…]
