### 2024-11-03 darrenjw/jax-smfsb#7 @darrenjw
Done (using black for now).
