### 2024-11-03 darrenjw/jax-smfsb#8 @cthoyt
I'd highly suggest running the following in the root of your repo before considering this issue closed:

```
ruff check --select N
```

More information about how PEP-8 naming checks works here: https://docs.astral.sh/ruff/rules/#pep8-naming-n. I guess it might be initially frustrating to get errors like

```
tests/test_spatial.py:65:5: N806 Variable `N` in function should be lowercase
   |
63 | def test_step_cle_2d():
64 |     M=16
65 |     N=20
   |     ^ N806
66 |     x0 = jnp.zeros((2,M,N))
67 |     lv = jsmfsb.models.lv()
```

but, people who read your code will highly appreciate not having cryptic variable names everywhere
