### 2024-11-03 5ed7774
add tests directory to formatter
