+++ tests/test_inference.py
@@ -7,107 +7,138 @@ import jax.numpy as jnp
 import jax.scipy as jsp
 
 
-
 def test_metropolis_hastings():
     key = jax.random.key(42)
-    data = jax.random.normal(key, 250)*2 + 5
-    llik = lambda k,x: jnp.sum(jsp.stats.norm.logpdf(data, x[0], x[1]))
-    prop = lambda k,x: jax.random.normal(k, 2)*0.1 + x
-    out = jsmfsb.metropolis_hastings(key, jnp.array([1.0,1.0]), llik, prop,
-                                   iters=1000, thin=2, verb=False)
-    assert(out.shape == (1000, 2))
+    data = jax.random.normal(key, 250) * 2 + 5
+    llik = lambda k, x: jnp.sum(jsp.stats.norm.logpdf(data, x[0], x[1]))
+    prop = lambda k, x: jax.random.normal(k, 2) * 0.1 + x
+    out = jsmfsb.metropolis_hastings(
+        key, jnp.array([1.0, 1.0]), llik, prop, iters=1000, thin=2, verb=False
+    )
+    assert out.shape == (1000, 2)
 
-    
+
 def test_pfmllik():
     def obsll(x, t, y, th):
-        return jnp.sum(jsp.stats.norm.logpdf((y-x)/10))
+        return jnp.sum(jsp.stats.norm.logpdf((y - x) / 10))
+
     def simX(k, t0, th):
         k1, k2 = jax.random.split(k)
-        return jnp.array([jax.random.poisson(k1, 50),
-                          jax.random.poisson(k2, 100)]).astype(jnp.float32)
+        return jnp.array(
+            [jax.random.poisson(k1, 50), jax.random.poisson(k2, 100)]
+        ).astype(jnp.float32)
+
     def step(k, x, t, dt, th):
         sf = jsmfsb.models.lv(th).step_cle()
         return sf(k, x, t, dt)
+
     mll = jsmfsb.pf_marginal_ll(50, simX, 0, step, obsll, jsmfsb.data.lv_noise_10)
     k = jax.random.key(42)
-    assert (mll(k, jnp.array([1, 0.005, 0.6])) > mll(k, jnp.array([2, 0.005, 0.6])))
+    assert mll(k, jnp.array([1, 0.005, 0.6])) > mll(k, jnp.array([2, 0.005, 0.6]))
 
 
 def test_abc_run():
     k0 = jax.random.key(42)
     k1, k2 = jax.random.split(k0)
-    data = jax.random.normal(k1, 250)*2 + 5
+    data = jax.random.normal(k1, 250) * 2 + 5
+
     def rpr(k):
-      return jnp.exp(jax.random.uniform(k, 2, minval=-3, maxval=3))
+        return jnp.exp(jax.random.uniform(k, 2, minval=-3, maxval=3))
+
     def rmod(k, th):
-      return jax.random.normal(k, 250)*th[1] + th[0]
+        return jax.random.normal(k, 250) * th[1] + th[0]
+
     def sumStats(dat):
-      return jnp.array([jnp.mean(dat), jnp.std(dat)])
+        return jnp.array([jnp.mean(dat), jnp.std(dat)])
+
     ssd = sumStats(data)
+
     def dist(ss):
-      diff = ss - ssd
-      return jnp.sqrt(jnp.sum(diff*diff))
+        diff = ss - ssd
+        return jnp.sqrt(jnp.sum(diff * diff))
+
     def rdis(k, th):
-      return dist(sumStats(rmod(k, th)))
+        return dist(sumStats(rmod(k, th)))
+
     p, d = jsmfsb.abc_run(k2, 100, rpr, rdis)
-    assert(len(p) == 100)
[…]
