### 2024-11-04 a2cc4c0
add docs to black formatting
