### 2024-11-05 2c55cab
fix ruff suggestions in the main source tree
