+++ src/jsmfsb/__init__.py
@@ -2,7 +2,7 @@ from jsmfsb.spn import *
 from jsmfsb.sim import *
 from jsmfsb.spatial import *
 from jsmfsb.inference import *
-from jsmfsb.smfsbSbml import *
+from jsmfsb.smfsb_sbml import *
 
 import jsmfsb.models
 import jsmfsb.data
