+++ src/jsmfsb/spn.py
@@ -70,7 +70,7 @@ class Spn:
             str(self.m),
         )
 
-    def step_gillespie(self, minHaz=1e-10, maxHaz=1e07):
+    def step_gillespie(self, min_haz=1e-10, max_haz=1e07):
         """Create a function for advancing the state of a SPN by using the
         Gillespie algorithm
 
@@ -81,9 +81,9 @@ class Spn:
 
         Parameters
         ----------
-        minHaz : float
+        min_haz : float
           Minimum hazard to consider before assuming 0. Defaults to 1e-10.
-        maxHaz : float
+        max_haz : float
           Maximum hazard to consider before assuming an explosion and
           bailing out. Defaults to 1e07.
 
@@ -105,8 +105,8 @@ class Spn:
         >>> stepLv = lv.step_gillespie()
         >>> stepLv(jax.random.key(42), lv.m, 0, 1)
         """
-        S = (self.post - self.pre).T
-        u, v = S.shape
+        sto = (self.post - self.pre).T
+        u, v = sto.shape
 
         @jit
         def advance(state):
@@ -114,10 +114,10 @@ class Spn:
             h = self.h(x, t)
             h0 = jnp.sum(h)
             key, k1, k2 = jax.random.split(key, 3)
-            t = jnp.where(h0 > maxHaz, 1e90, t)
-            t = jnp.where(h0 < minHaz, 1e90, t + jax.random.exponential(k1) / h0)
+            t = jnp.where(h0 > max_haz, 1e90, t)
+            t = jnp.where(h0 < min_haz, 1e90, t + jax.random.exponential(k1) / h0)
             j = jax.random.choice(k2, v, p=h / h0)
-            xn = jnp.add(x, S[:, j])
+            xn = jnp.add(x, sto[:, j])
             return (key, x, xn, t)
 
         @jit
@@ -167,8 +167,8 @@ class Spn:
         >>> k = jax.random.key(42)
         >>> stepLv(k, lv.m, 0, 1)
         """
-        S = (self.post - self.pre).T
-        u, v = S.shape
+        sto = (self.post - self.pre).T
+        u, v = sto.shape
 
         @jit
         def advance(state):
@@ -176,7 +176,7 @@ class Spn:
             key, k1 = jax.random.split(key)
             h = self.h(x, t)
             r = jax.random.poisson(k1, h * dt)
-            x = jnp.add(x, S.dot(r))
+            x = jnp.add(x, sto.dot(r))
             # TODO: sort out negative values
             # x = jnp.where(x < 0, -x, x)
             t = t + dt
@@ -228,14 +228,14 @@ class Spn:
         >>> k = jax.random.key(42)
         >>> stepLv(k, lv.m, 0, 1)
         """
-        S = (self.post - self.pre).T
+        sto = (self.post - self.pre).T
 
         @jit
         def advance(state):
             key, x, t = state
             key, k1 = jax.random.split(key)
             h = self.h(x, t)
-            x = jnp.add(x, S.dot(h * dt))
+            x = jnp.add(x, sto.dot(h * dt))
             x = jnp.where(x < 0, -x, x)
             t = t + dt
             return (key, x, t)
@@ -286,8 +286,8 @@ class Spn:
         >>> stepLv = lv.step_cle(0.001)
         >>> stepLv(jax.random.key(42), lv.m, 0, 1)
         """
-        S = (self.post - self.pre).T
-        v = S.shape[1]
+        sto = (self.post - self.pre).T
+        v = sto.shape[1]
         sdt = np.sqrt(dt)
 
         @jit
@@ -296,7 +296,7 @@ class Spn:
             key, k1 = jax.random.split(key)
             h = self.h(x, t)
[…]
