+++ src/jsmfsb/spatial.py
@@ -7,7 +7,7 @@ import jax.numpy as jnp
 import jax.lax as jl
 
 
-def sim_time_series_1d(key, x0, t0, tt, dt, stepFun, verb=False):
+def sim_time_series_1d(key, x0, t0, tt, dt, step_fun, verb=False):
     """Simulate a model on a regular grid of times, using a function (closure)
     for advancing the state of the model
 
@@ -32,7 +32,7 @@ def sim_time_series_1d(key, x0, t0, tt, dt, stepFun, verb=False):
       The time step of the output. Note that this time step relates
       only to the recorded output, and has no bearing on the
       accuracy of the simulation process.
-    stepFun : function
+    step_fun : function
       A function (closure) for advancing the state of the process,
       such as produced by `step_gillespie_1d`.
     verb : boolean
@@ -57,16 +57,16 @@ def sim_time_series_1d(key, x0, t0, tt, dt, stepFun, verb=False):
     >>> k0 = jax.random.key(42)
     >>> jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, stepLv1d, True)
     """
-    N = int((tt - t0) // dt + 1)
+    nn = int((tt - t0) // dt + 1)
     u, n = x0.shape
-    keys = jax.random.split(key, N)
+    keys = jax.random.split(key, nn)
 
     @jit
     def advance(state, key):
         x, t = state
-        if verb == True:
+        if verb:
             jax.debug.print("{t}", t=t)
-        x = stepFun(key, x, t, dt)
+        x = step_fun(key, x, t, dt)
         t = t + dt
         return (x, t), x
 
@@ -74,7 +74,7 @@ def sim_time_series_1d(key, x0, t0, tt, dt, stepFun, verb=False):
     return jnp.moveaxis(arr, 0, 2)
 
 
-def sim_time_series_2d(key, x0, t0, tt, dt, stepFun, verb=False):
+def sim_time_series_2d(key, x0, t0, tt, dt, step_fun, verb=False):
     """Simulate a model on a regular grid of times, using a function (closure)
     for advancing the state of the model
 
@@ -99,7 +99,7 @@ def sim_time_series_2d(key, x0, t0, tt, dt, stepFun, verb=False):
       The time step of the output. Note that this time step relates
       only to the recorded output, and has no bearing on the
       accuracy of the simulation process.
-    stepFun : function
+    step_fun : function
       A function (closure) for advancing the state of the process,
       such as produced by `step_gillespie_2d`.
     verb : boolean
@@ -125,16 +125,16 @@ def sim_time_series_2d(key, x0, t0, tt, dt, stepFun, verb=False):
     >>> k0 = jax.random.key(42)
     >>> jsmfsb.sim_time_series_2d(k0, x0, 0, T, 1, stepLv2d, True)
     """
-    N = int((tt - t0) // dt + 1)
+    nn = int((tt - t0) // dt + 1)
     u, m, n = x0.shape
-    keys = jax.random.split(key, N)
+    keys = jax.random.split(key, nn)
 
     @jit
     def advance(state, key):
         x, t = state
-        if verb == True:
+        if verb:
             jax.debug.print("{t}", t=t)
-        x = stepFun(key, x, t, dt)
+        x = step_fun(key, x, t, dt)
         t = t + dt
         return (x, t), x
 
