### 2024-11-12 e46247d
better introductory documentation
