+++ README.md
@@ -20,7 +20,22 @@ pip install --upgrade jsmfsb
 **Note** that a number of breaking syntax changes (more pythonic names) were introduced in version 1.1.0. If you upgrade to a version >= 1.1.0 from a version prior to 1.1.0 you will have to update syntax to the new style.
 
 
-You can test that your installation is working with the following example.
+You can test that your installation is working by entering the following at a python prompt:
+
+```python
+import jax
+import jsmfsb
+```
+If these both return silently, you are probably good to go.
+
+## Basic usage
+
+Note that **the book**, and its associated [github repo](https://github.com/darrenjw/smfsb) is the main source of documentation for this library. The code in the book is in R, but the code in this library is supposed to mirror the R code, but in Python.
+
+### Using a model built-in to the library
+
+First, see how to simulate a built-in (Lotka-Volterra predator-prey) model:
+
 ```python
 import jax
 import jsmfsb
@@ -43,7 +58,86 @@ axis.legend(lvmod.n)
 fig.savefig("lv.pdf")
 ```
 
+Standard python docstring documentation is available. Usage information can be obtained from the python REPL with commands like `help(jsmfsb.Spn)`, `help(jsmfsb.Spn.step_gillespie)` or `help(jsmfsb.sim_time_series)`. This documentation is also available on [ReadTheDocs](https://jax-smfsb.readthedocs.io/). The API documentation contains very minimal usage examples. For more interesting examples, see the [demos directory](https://github.com/darrenjw/jax-smfsb/tree/main/demos).
+
+### Creating and simulating a model
+
+Next, let's create and simulate our own (SIR epidemic) model by specifying a stochastic Petri net explicitly.
+```python
+import jax.numpy as jnp
+sir = jsmfsb.Spn(["S", "I", "R"], ["S->I", "I->R"],
+	[[1,1,0],[0,1,0]], [[0,2,0],[0,0,1]],
+	lambda x, t: jnp.array([0.3*x[0]*x[1]/200, 0.1*x[1]]),
+	[197.0, 3, 0])
+stepSir = sir.step_poisson()
+sample = jsmfsb.sim_sample(k0, 500, sir.m, 0, 20, stepSir)
+fig, axis = plt.subplots()
+axis.hist(sample[:,1],30)
+axis.set_title("Infected at time 20")
+plt.savefig("sIr.pdf")
+```
+
+### Reading and parsing models in SBML and SBML-shorthand
+
+Note that you can read in SBML or SBML-shorthand models that have been designed for discrete stochastic simulation into a stochastic Petri net directly. To read and parse an SBML model, use
+```python
+m = jsmfsb.file_to_spn("myModel.xml")
+```
+Note that if you are working with SBML models in Python using [libsbml](https://pypi.org/project/python-libsbml/), then there is also a function `model_to_spn` which takes a libsbml model object.
+
+To read and parse an SBML-shorthand model, use
+```python
+m = jsmfsb.mod_to_spn("myModel.mod")
+```
+There is also a function `shorthand_to_spn` which expects a python string containing a shorthand model. This is convenient for embedding shorthand models inside python scripts, and is particularly convenient when working with things like Jupyter notebooks. Below follows a complete session to illustrate the idea by creating and simulating a realisation from a discrete stochastic SEIR model.
+```python
+import jax
+import jsmfsb
+import jax.numpy as jnp
+
+seirSH = """
+@model:3.1.1=SEIR "SEIR Epidemic model"
+ s=item, t=second, v=litre, e=item
+@compartments
+ Pop
[…]
