### 2024-11-16 fc08976
tidy up the demos (ruff-approved variable names)
