+++ demos/pf_marginal_ll.py
@@ -11,7 +11,7 @@ def obsll(x, t, y, th):
     return jnp.sum(jsp.stats.norm.logpdf(y - x, scale=10))
 
 
-def simX(k, t0, th):
+def sim_x(k, t0, th):
     k1, k2 = jax.random.split(k)
     return jnp.array([jax.random.poisson(k1, 50), jax.random.poisson(k2, 100)]).astype(
         jnp.float32
@@ -24,7 +24,7 @@ def step(k, x, t, dt, th):
     return sf(k, x, t, dt)
 
 
-mll = jsmfsb.pf_marginal_ll(100, simX, 0, step, obsll, jsmfsb.data.lv_noise_10)
+mll = jsmfsb.pf_marginal_ll(100, sim_x, 0, step, obsll, jsmfsb.data.lv_noise_10)
 
 k = jax.random.split(jax.random.key(42), 5)
 print(mll(k[0], jnp.array([1, 0.005, 0.6])))
