+++ demos/step_cle_1d.py
@@ -11,11 +11,11 @@ T = 30
 x0 = jnp.zeros((2, N))
 lv = jsmfsb.models.lv()
 x0 = x0.at[:, int(N / 2)].set(lv.m)
-stepLv1d = lv.step_cle_1d(jnp.array([0.6, 0.6]))
+step_lv_1d = lv.step_cle_1d(jnp.array([0.6, 0.6]))
 k0 = jax.random.key(42)
-x1 = stepLv1d(k0, x0, 0, 1)
+x1 = step_lv_1d(k0, x0, 0, 1)
 print(x1)
-out = jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, stepLv1d, True)
+out = jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, step_lv_1d, True)
 # print(out)
 
 fig, axis = plt.subplots()
