+++ demos/step_gillespie_1d.py
@@ -12,10 +12,10 @@ x0 = jnp.zeros((2, N))
 lv = jsmfsb.models.lv()
 x0 = x0.at[:, int(N / 2)].set(lv.m)
 k0 = jax.random.key(42)
-stepLv1d = lv.step_gillespie_1d(jnp.array([0.6, 0.6]))
-x1 = stepLv1d(k0, x0, 0, 1)
+step_lv_1d = lv.step_gillespie_1d(jnp.array([0.6, 0.6]))
+x1 = step_lv_1d(k0, x0, 0, 1)
 print(x1)
-out = jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, stepLv1d, True)
+out = jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, step_lv_1d, True)
 # print(out)
 
 fig, axis = plt.subplots()
