### 2024-11-16 2e98e6a
apply some ruff fixes
