+++ tests/test_spatial.py
@@ -4,7 +4,6 @@
 import jsmfsb
 import jax
 import jax.numpy as jnp
-import matplotlib.pyplot as plt
 import jsmfsb.models
 
 
