+++ demos/m-lv-cle.py
@@ -3,10 +3,6 @@
 # use a pre-defined model
 
 import jax
-import jax.numpy as jnp
-import jax.scipy as jsp
-import jax.lax as jl
-from jax import grad, jit
 
 import jsmfsb
 
