### 2024-11-16 a32285c
more ruff fixes
