+++ demos/metropolis_hastings.py
@@ -5,16 +5,19 @@ import jsmfsb
 import jax
 import jax.numpy as jnp
 import jax.scipy as jsp
+import matplotlib.pyplot as plt
 
 k0 = jax.random.key(42)
 k1, k2 = jax.random.split(k0)
 data = jax.random.normal(k1, 250) * 2 + 5
-llik = lambda k, x: jnp.sum(jsp.stats.norm.logpdf(data, x[0], x[1]))
-prop = lambda k, x: jax.random.normal(k, 2) * 0.1 + x
+
+def llik(k, x):
+    return jnp.sum(jsp.stats.norm.logpdf(data, x[0], x[1]))
+
+def prop(k, x):
+    return jax.random.normal(k, 2) * 0.1 + x
+
 postmat = jsmfsb.metropolis_hastings(k2, jnp.array([1.0, 1.0]), llik, prop, verb=False)
-
-
-import matplotlib.pyplot as plt
 
 fig, axes = plt.subplots(3, 2)
 axes[0, 0].scatter(postmat[:, 0], postmat[:, 1], s=0.5)
