+++ demos/metropolis_hastings.py
@@ -11,11 +11,14 @@ k0 = jax.random.key(42)
 k1, k2 = jax.random.split(k0)
 data = jax.random.normal(k1, 250) * 2 + 5
 
+
 def llik(k, x):
     return jnp.sum(jsp.stats.norm.logpdf(data, x[0], x[1]))
 
+
 def prop(k, x):
     return jax.random.normal(k, 2) * 0.1 + x
+
 
 postmat = jsmfsb.metropolis_hastings(k2, jnp.array([1.0, 1.0]), llik, prop, verb=False)
 
