### 2024-11-16 34bc0b1
tidy readme examples
