### 2024-11-16 19527de
add a tutorial template
