+++ docs/source/tutorial.rst
@@ -1,0 +1,5 @@
+# The jax-smfsb tutorial
+
+Blah
+
+
