### 2024-11-16 3c3976b
fix link to tutorial template
