+++ docs/source/tutorial.rst
@@ -1,7 +1,213 @@
 The jax-smfsb tutorial
 ----------------------
 
-Blah
+This tutorial assumes that the package has already been installed, following the instructions in the `package readme <https://pypi.org/project/jsmfsb/>`__.
+
+We begin with non-spatial stochastic simulation.
+
+Non-spatial simulation
+----------------------
+
+Using a model built-in to the library
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+First, see how to simulate a built-in (Lotka-Volterra predator-prey)
+model:
+
+.. code:: python
+
+   import jax
+   import jsmfsb
+
+   lvmod = jsmfsb.models.lv()
+   step = lvmod.step_gillespie()
+   k0 = jax.random.key(42)
+   out = jsmfsb.sim_time_series(k0, lvmod.m, 0, 30, 0.1, step)
+   assert(out.shape == (300, 2))
+
+Here we used the ``lv`` model. Other built-in models include ``id`` (immigration-death), ``bd`` (birth-death), ``dimer`` (dimerisation kinetics), ``mm`` (Michaelis-Menten enzyme kinetics) and ``sir`` (SIR epdiemic model). The models are of class ``Spn`` (stochastic Petri net), the main data type used in the package. Note the use of the ``step_gillespie`` method, defined on all ``Spn`` models, which returns a function for simulating from the transition kernel of the model, using the Gillespie algorithm. This function can be used with the ``sim_time_series`` function for simulating model trajectories on a regular time grid. Note that all stochastic simulation functions in this package take a `JAX random number key <https://jax.readthedocs.io/en/latest/random-numbers.html>`__ as their first argument. Alternative simulation algorithms include ``step_poisson`` (Poisson time-stepping), ``step_cle`` (Euler-Maruyama simulation from the associated chemical Langevin equation) and ``step_euler`` (Euler simulation from the continuous deterministic approximation to the model).
+
+If you have ``matplotlib`` installed (``pip install matplotlib``), then
+you can also plot the results with:
+
+.. code:: python
+
+   import matplotlib.pyplot as plt
+   fig, axis = plt.subplots()
+   for i in range(2):
+       axis.plot(range(out.shape[0]), out[:,i])
+
+   axis.legend(lvmod.n)
+   fig.savefig("lv.pdf")
+
+Standard python docstring documentation is available. Usage information
+can be obtained from the python REPL with commands like
+``help(jsmfsb.Spn)``, ``help(jsmfsb.Spn.step_gillespie)`` or
+``help(jsmfsb.sim_time_series)``. This documentation is also available
+on `ReadTheDocs <https://jax-smfsb.readthedocs.io/>`__. The API
+documentation contains minimal usage examples.
+
+Creating and simulating a model
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+Next, let’s create and simulate our own (SIR epidemic) model by
+specifying a stochastic Petri net ``Spn`` object explicitly. We must provide species and reaction names, stoichiometry matrices, reaction rates and initial conditions. This time we use approximate Poisson simulation rather than exact simulation via the Gillespie algorithm.
+
+.. code:: python
+
+   import jax.numpy as jnp
+   sir = jsmfsb.Spn(["S", "I", "R"], ["S->I", "I->R"],
+       [[1,1,0], [0,1,0]], [[0,2,0], [0,0,1]],
+       lambda x, t: jnp.array([0.3*x[0]*x[1]/200, 0.1*x[1]]),
[…]
