+++ docs/source/tutorial.rst
@@ -152,24 +152,290 @@ In addition to methods such as ``step_gillespie`` and ``step_cle`` for well-mixe
 ~~~~~~~~~~~~~
 
 
+.. code:: python
+
+   import jsmfsb
+   import jax
+   import jax.numpy as jnp
+   import matplotlib.pyplot as plt
+   import jsmfsb.models
+
+   N = 20
+   T = 30
+   x0 = jnp.zeros((2, N))
+   lv = jsmfsb.models.lv()
+   x0 = x0.at[:, int(N / 2)].set(lv.m)
+   k0 = jax.random.key(42)
+   step_lv_1d = lv.step_gillespie_1d(jnp.array([0.6, 0.6]))
+   x1 = step_lv_1d(k0, x0, 0, 1)
+   print(x1)
+   out = jsmfsb.sim_time_series_1d(k0, x0, 0, T, 1, step_lv_1d, True)
+   # print(out)
+
+   fig, axis = plt.subplots()
+   for i in range(2):
+       axis.imshow(out[i, :, :])
+       axis.set_title(lv.n[i])
+       fig.savefig(f"step_gillespie_1d{i}.pdf")
+
+
+
+
 2d simulation
 ~~~~~~~~~~~~~
 
 
 
+.. code:: python
+
+   import jsmfsb
+   import jax
+   import jax.numpy as jnp
+   import matplotlib.pyplot as plt
+   import jsmfsb.models
+
+   M = 200
+   N = 250
+   T = 30
+   x0 = jnp.zeros((2, M, N))
+   lv = jsmfsb.models.lv()
+   x0 = x0.at[:, int(M / 2), int(N / 2)].set(lv.m)
+   step_lv_2d = lv.step_cle_2d(jnp.array([0.6, 0.6]), 0.1)
+   k0 = jax.random.key(42)
+   x1 = step_lv_2d(k0, x0, 0, T)
+
+   fig, axis = plt.subplots()
+   for i in range(2):
+       axis.imshow(x1[i, :, :])
+       axis.set_title(lv.n[i])
+       fig.savefig(f"step_cle_2df{i}.pdf")
+
+
+    
+
 Bayesian parameter inference
 ----------------------------
 
+
+
 ABC
 ~~~
+
+.. code:: python
+
+   import jsmfsb
+   import jax
+   import jax.numpy as jnp
+   import matplotlib.pyplot as plt
+
+   data = jsmfsb.data.lv_perfect[:, 1:3]
+
+   def rpr(k):
+       k1, k2, k3 = jax.random.split(k, 3)
+       return jnp.exp(
+	   jnp.array(
+	       [
+		   jax.random.uniform(k1, minval=-3, maxval=3),
+		   jax.random.uniform(k2, minval=-8, maxval=-2),
+		   jax.random.uniform(k3, minval=-4, maxval=2),
+	       ]
+	   )
+       )
+
+   def rmod(k, th):
+       return jsmfsb.sim_time_series(
+	   k, jnp.array([50.0, 100.0]), 0, 30, 2, jsmfsb.models.lv(th).step_cle(0.1)
+       )
+
+   def sum_stats(dat):
+       return dat
+
+   ssd = sum_stats(data)
+
+   def dist(ss):
+       diff = ss - ssd
+       return jnp.sqrt(jnp.sum(diff * diff))
+
+   def rdis(k, th):
+       return dist(sum_stats(rmod(k, th)))
+
+   k0 = jax.random.key(42)
+   p, d = jsmfsb.abc_run(k0, 1000000, rpr, rdis, batch_size=100000, verb=False)
+
[…]
