### 2024-11-17 536a0b6
link to the tutorial and clean up
