### 2024-11-17 darrenjw/jax-smfsb#8 @darrenjw
Have had a good go at cleaning up ruff suggestions
