### 2024-11-17 darrenjw/jax-smfsb#9 @darrenjw
I've added a tutorial document for ReadTheDocs, and a RTD badge.
