### 2024-11-17 #7329 @darrenjw
The suggestions from @cthoyt are all good - just time-consuming. I've had a reasonable go at addressing most of them, so I think it's probably fine to start the real review process now. Thanks,
