### 2024-11-18 #7491 @editorialbot
[REVIEW]: jax-smfsb: A python library for stochastic systems biology modelling and inference
**Submitting author:** @darrenjw (<a href="http://orcid.org/0000-0003-0736-802X">Darren James Wilkinson</a>)
**Repository:** https://github.com/darrenjw/jax-smfsb
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.5
