### 2024-12-02 #7491 @jromanowska
@antoineruzette, @cthoyt - could you generate your checklists and start the review?
