### 2024-12-28 9e72825
bump numpy version to >= 2
