+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = "README.md"
 requires-python = ">=3.8"
 dependencies = [
   "sbmlsh",
-  "numpy<2",
+  "numpy>=2",
   "scipy",
   "jaxlib"
 ]
