+++ .github/workflows/ci.yml
@@ -26,4 +26,4 @@ jobs:
 
       - name: Run tests
         run: |
-          uvx -p 3.11 --with jsmfsb pytest
+          uvx -p 3.11 --with jax --with jsmfsb pytest
