+++ README.md
@@ -1,6 +1,8 @@
 # JAX-SMfSB (jsmfsb)
 
-[![](https://readthedocs.org/projects/jax-smfsb/badge/)](https://jax-smfsb.readthedocs.io/en/latest/index.html) [![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
+[![](https://readthedocs.org/projects/jax-smfsb/badge/)](https://jax-smfsb.readthedocs.io/en/latest/index.html)
+[![](https://github.com/darrenjw/python-smfsb/actions/workflows/ci.yml/badge.svg)](https://github.com/darrenjw/python-smfsb/actions)
+[![Code style: black](https://img.shields.io/badge/code%20style-black-000000.svg)](https://github.com/psf/black)
 
 ## SMfSB code in Python+JAX
 
