### 2024-12-30 aa9fab0
batch size for timed cle sim
