### 2025-01-02 44db802
Identified bug in sim_time_series. Will add tests to highlight the bug before fixing.
