### 2025-01-02 3fbddd4
added tests for correct behaviour of sim_time_series (currently failing)
