### 2025-01-02 019e754
bug fixed and initial tests passing
