### 2025-01-02 4f2430d
speed up dsmts runner by replacing a for loop with a scan
