+++ dsmts/test_models.py
@@ -4,7 +4,7 @@
 import dsmts
 import jax.numpy as jnp
 
-N = 1000  # TODO: bump back up to 10000?
+N = 10000
 
 
 def check_model(file_stem):
