### 2025-01-22 darrenjw/jax-smfsb#11 @zeyus
"RuntimeWarning: overflow encountered in cast" generated in demo code
I'm currently reviewing the [`jax-smfsb` submission for the JOSS](https://github.com/openjournals/joss-reviews/issues/7491), and while running the examples in the documentation, I encountered the following warning (without a visible backtrace):

```text
G:\envs\mamba\joss\Lib\site-packages\jax\_src\interpreters\xla.py:135: RuntimeWarning: overflow encountered in cast
  return np.asarray(
```

This warning does not seem to affect the output, and generated plots in other examples that also produce the warning seem sane.

For example in the [Lotka-Volterra predator-prey model](https://jax-smfsb.readthedocs.io/en/latest/source/tutorial.html#using-a-model-built-in-to-the-library)

```python
import jax
import jsmfsb

lvmod = jsmfsb.models.lv()
step = lvmod.step_gillespie()
k0 = jax.random.key(42)
out = jsmfsb.sim_time_series(k0, lvmod.m, 0, 30, 0.1, step)
assert(out.shape == (300, 2))
```

The warning is generated during the `sim_time_series` call.

I have tried wrapping it in a method and using the `@jax.jit` decorator to see if I could get more information, but that actually makes the warning disappear.

This occurs both when using the version of `jax-smfsb` from PyPi, as well as using the wheel generated from `hatch build`

### Environment information

OS: Windows 11 Pro 10.0.26120 Build 26120
CPU: Intel Core i5-4690K
RAM: 24 GB
GPU (not available in jax on windows): NVIDIA GTX 1070 8GB

Output of `python --version`:

```text
Python 3.13.1
```

Output of `pip --list` (from inside a mamba environment):

```text
Package         Version
--------------- -----------
contourpy       1.3.1
cycler          0.12.1
fonttools       4.55.4
jax             0.5.0
jaxlib          0.5.0
jsmfsb          1.1.3
kiwisolver      1.4.8
matplotlib      3.10.0
ml_dtypes       0.5.1
numpy           2.2.2
opt_einsum      3.4.0
packaging       24.2
pillow          11.1.0
pip             24.3.1
pyparsing       3.2.1
python-dateutil 2.9.0.post0
python-libsbml  5.20.4
sbmlsh          0.0.3
scipy           1.15.1
six             1.17.0
```

Let me know if there's an obvious way to help debug this. I'm not experienced with jax.
