### 2025-01-22 #7491 @zeyus
@jromanowska So far so good. I would just like some input on the following two items:

1) Regarding **Functionality: Performance** I'm currently testing this on Windows (there package meta specifies OS independent) but jax support is [CPU only on Windows x86_64](https://jax.readthedocs.io/en/latest/installation.html#supported-platforms) and does not support GPU meaning:
    - I am unable to accurately assess the claim regarding two-fold performance increase; but
    - I can run the package, and demos, and they complete in a short amount of time and produce results

    As such, I'm wondering how far to extend the testing, (i.e. on WSL2 with CUDA)? I don't have any other PC with a GPU, only an office Macbook Pro (Apple silicon, OSX) and a personal laptop (Arch Linux, Intel CPU). I have no reason to doubt the claim in any case, if the operations run on the GPU then there should be a significant performance increase (the package uses native jax methods for e.g. MCMC inference).

2) While testing some of the examples, I have experienced a warning about an overflow, the examples complete successfully (at least the generated plots look reasonable), but it indicates an issue somewhere in the stack (though it may not even originate with this library). I'm not familiar with jax, so I cannot estimate how likely this warning is to lead to unexpected output, so I'm wondering what to do with this information? I have created a Github issue (https://github.com/darrenjw/jax-smfsb/issues/11) so the author is aware and can maybe comment on if it's something that needs to be addressed (or is even within the scope of the package itself)

Thanks.
