### 2025-01-22 2d553fc
Added jax as a hatch test extra-dependency.
