### 2025-01-22 87b2e68
fix for Issue #11 raised by @zeyus
