### 2025-01-22 darrenjw/jax-smfsb#11 @darrenjw
Thanks for raising this. It affected most (if not all) platforms. I'd been ignoring it, as I'd checked carefully that it wasn't affecting the correctness of results. But I've now tracked down the source and fixed it, and pushed an update to PyPI. Please check and close if you are happy.
