### 2025-01-22 #7491 @darrenjw
Perhaps I can give some additional information regarding the two points raised above.
1. I try not to make too strong statements about performance, as it's very hardware dependent. For example, the machine I'm using right now has a very fast CPU with 20 cores but a low end consumer gaming GPU. On this machine running on the GPU is typically _slower_ than running on the CPU. But if you have access to a high-end GPU it is a different story. The main performance claims I make relate to the speed of this package (jsmfsb) relative to the non-JAX version (smfsb). The performance difference is consistently one to three orders of magnitude for computationally intensive tasks when running on decent hardware (even without a GPU). Anyone interested can easily verify this by running the demos for the two packages, which are largely direct translations of each other.
2. I had checked that the overflow warning was not affecting the correctness of results, but in any case I think I've now fixed it, as discussed on the issue on the repo.
