### 2025-01-23 #7491 @jromanowska
Thank you, @zeyus, for a speedy review 🥳 , and @darrenjw, for the replies and fix 🧰 .

As for the performance, you don't have to run it on another hardware, but if it's simple to test against the [smfb](https://pypi.org/project/smfsb/) package, you could try that.
