### 2025-01-23 #7491 @zeyus
@darrenjw That's completely fair, and it seems logical to me, I just wasn't sure how in depth to go with testing. As an aside I'm currently trying to wrap my head around how to implement Bayesian inference GPU optimizations myself (for a hierarchical cognitive modelling framework) implemented as a network, so it's interesting to me.

Regarding the overflow warning, I also didn't see anything detrimental in the results, but nice work fixing it, thanks!

@jromanowska @darrenjw I just compared a few of the example models, and indeed as soon as it is even slightly more complex (e.g. the LV predator-prey simulation) there's a clear performance lead (statistically significant according to `scipy.stats.ttest_rel` t: 112.5997, p: 1.7449e-15, df: 9) in `jsmfsb` over `smfsb`.

| Run | smfsb (s) | jsmfsb (s) |
| --- | --- | --- |
| 0 | 23.1623 | 0.6646 |
| 1 | 24.4116 | 0.4355 |
| 2 | 24.0472 | 0.5670 |
| 3 | 23.5138 | 0.4270 |
| 4 | 25.1928 | 0.4485 |
| 5 | 23.8482 | 0.4140 |
| 6 | 23.2873 | 0.4112 |
| 7 | 24.2407 | 0.4158 |
| 8 | 24.4115 | 0.4314 |
| 9 | 23.4845 | 0.4720 |
| **Mean** | 23.9600 | 0.4687 |
| **Std** | 0.5950 | 0.0786 |

So, all good from my end! :)
