### 2025-02-03 #7491 @benlansdell
Hi all, apologies for the delay on this. I've completed my review. Everything looks good - this all looks very well-written, functional and well documented. I have very minor comments around some of the doc strings and performance claims in the paper.

First, the documentation: I notice some of the examples in the doc strings (e.g. https://jax-smfsb.readthedocs.io/en/latest/source/jsmfsb.html#jsmfsb.models.id, https://jax-smfsb.readthedocs.io/en/latest/source/jsmfsb.html#jsmfsb.inference.abc_run, and a few others) seem to reference the package `smfsb` (the non-Jax version). I don't see this as a dependency, and assume it is meant to be `jsmfsb` (the code worked when I replaced it with that). Perhaps these need to be updated?

And in the paper, continuing the discussion on performance: as performance is a key motivation as outlined in the statement of need, it may be nice to provide one or two specific benchmarks, exactly along the lines of what @zeyus did in this thread. Either in the `paper.md` or as a demo in the repo. The paper does state that timings are hardware dependent, which is fair, though I do think that two orders of magnitude is a significant enough speed up to warrant drawing more attention to, even with the caveat you mention. Right now this result is the very last sentence of the paper.

It's not clear, for some of the features in the first paragraph of the `Features` section, which are new to `jax-smfsb`. Are all these features new? (Presumably the ones to do with JAX arrays are...)
