### 2025-02-03 c8d3efe
fix some docstring errors, highlighted by @benlansdell
