### 2025-02-03 eda9c17
some clarifications, as suggested by @benlansdell
