+++ paper/paper.md
@@ -38,9 +38,9 @@ In addition to describing the mathematical framework for stochastic modelling, s
 
 # Features
 
-In addition to providing a small library of pre-defined models, and allowing direct specification of models as stochastic Petri nets using JAX arrays, `jax-smfsb` can also parse models encoded in the Systems Biology Markup Language (SBML) [@SBML] using `libSBML` [@Bornstein2008], in addition to those encoded using the SBML-shorthand notation used in @Wilkinson2018. Exact and approximate stochastic simulation algorithms are provided for both the well-mixed and spatial (reaction-diffusion) case. Exact and approximate Bayesian inference algorithms for parameter inference are provided based on ABC [@Marin2011], ABC-SMC [@Toni2008] or particle MCMC [@Andrieu2010] approaches.
+Similarly to the NumPy version of the library, in addition to providing a small library of pre-defined models, and allowing direct specification of models as stochastic Petri nets using arrays, `jax-smfsb` can also parse models encoded in the Systems Biology Markup Language (SBML) [@SBML] using `libSBML` [@Bornstein2008], in addition to those encoded using the SBML-shorthand notation used in @Wilkinson2018. Exact and approximate stochastic simulation algorithms are provided for both the well-mixed and spatial (reaction-diffusion) case. Exact and approximate Bayesian inference algorithms for parameter inference are provided based on ABC [@Marin2011], ABC-SMC [@Toni2008] or particle MCMC [@Andrieu2010] approaches.
 
-Many demos are provided in the `demos` directory. These are direct equivalents to the demos provided in the `python-smfsb` library, so it is easy to compare the performance of the Python+JAX implementations to the more conventional Python+Numpy implementation. Precise timings are very hardware dependent, but on a high-specification machine, speed-ups of around two orders of magnitude for computationally intensive simulation or inference tasks can be expected.
+Many demos are provided in the `demos` directory. These are direct equivalents to the demos provided in the `python-smfsb` library, so it is easy to compare the performance of the Python+JAX implementations to the more conventional Python+Numpy implementation. Precise timings are very hardware and problem dependent, but on a high-specification machine, speed-ups of around two orders of magnitude for computationally intensive simulation or inference tasks can be expected. On a Linux server with an Intel i7-12700 processor, the CPU-only version of `jax-smfsb` gives speedup factors relative to `python-smfsb` ranging from around 50 to around 2000 on the timing examples included in the `demos` directory.
 
 # References
 
