### 2025-02-03 #7491 @darrenjw
Thanks for the review, @benlansdell

I've fixed the docstrings that you highlighted, and have tried to systematically search for and fix similar issues.

Regarding performance, I'm still a little reluctant to get too specific, as the speedups are very problem dependent as well as hardware dependent. However, I've added an extra sentence at the end of the paper with some example speedups - see what you think.

Actually, the features in jax-smfsb are all essentially present in python-smfsb - I've now tried to clarify this.
