### 2025-02-04 #7491 @benlansdell
@darrenjw Great, thanks for fixing up the docstrings. And, yes, I think that's a fair additional comment about the timing you added.

All looks good from me now.
