### 2025-02-04 #7491 @jromanowska
@darrenjw, could you complete the **Additional Author Tasks** mentioned above?
