### 2025-02-04 #7491 @jromanowska
@darrenjw, please correct some references:

- CRAN packages have now DOIs, and the editorialbot has found a DOI for your R package {smfsb}, please add this one
- titles "Particle markov chain monte carlo methods" and "Bayesian parameter inference for stochastic biochemical network models using particle markov chain monte carlo" should probably have capital letters in "Markov" and "Monte Carlo"?
- JAX has now slightly [updated citation](https://github.com/jax-ml/jax?tab=readme-ov-file#citing-jax)
- title "Approximate bayesian computation scheme for parameter inference and model selection" needs capitalized "Bayesian"
- add the DOI for your book :) 10.1201/9781351000918
