### 2025-02-04 5e88965
fix paper references for JOSS
