+++ paper/paper.bib
@@ -1,16 +1,16 @@
 
-@article{Andrieu2010, title={Particle Markov Chain Monte Carlo Methods}, volume=72, ISSN={1467-9868}, url={http://dx.doi.org/10.1111/j.1467-9868.2009.00736.x}, DOI={10.1111/j.1467-9868.2009.00736.x}, number=3, journal={Journal of the Royal Statistical Society Series B: Statistical Methodology}, publisher={Oxford University Press (OUP)}, author={Andrieu, Christophe and Doucet, Arnaud and Holenstein, Roman}, year=2010, month=may, pages={269–342} }
+@article{Andrieu2010, title={Particle {M}arkov Chain {M}onte {C}arlo Methods}, volume=72, ISSN={1467-9868}, url={http://dx.doi.org/10.1111/j.1467-9868.2009.00736.x}, DOI={10.1111/j.1467-9868.2009.00736.x}, number=3, journal={Journal of the Royal Statistical Society Series B: Statistical Methodology}, publisher={Oxford University Press (OUP)}, author={Andrieu, Christophe and Doucet, Arnaud and Holenstein, Roman}, year=2010, month=may, pages={269–342} }
 
 @article{Bezgin2023, title={JAX-Fluids: A fully-differentiable high-order computational fluid dynamics solver for compressible two-phase flows}, volume={282}, ISSN={0010-4655}, url={http://dx.doi.org/10.1016/j.cpc.2022.108527}, DOI={10.1016/j.cpc.2022.108527}, journal={Computer Physics Communications}, publisher={Elsevier BV}, author={Bezgin, Deniz A. and Buhendwa, Aaron B. and Adams, Nikolaus A.}, year={2023}, month=jan, pages={108527} }
 
 @article{Bornstein2008, title={LibSBML: an API Library for SBML}, volume={24}, ISSN={1367-4803}, url={http://dx.doi.org/10.1093/bioinformatics/btn051}, DOI={10.1093/bioinformatics/btn051}, number={6}, journal={Bioinformatics}, publisher={Oxford University Press (OUP)}, author={Bornstein, Benjamin J. and Keating, Sarah M. and Jouraku, Akiya and Hucka, Michael}, year={2008}, month=feb, pages={880–881} }
 
-@article{Golightly2011, title={Bayesian parameter inference for stochastic biochemical network models using particle Markov chain Monte Carlo}, volume=1, ISSN={2042-8901}, url={http://dx.doi.org/10.1098/rsfs.2011.0047}, DOI={10.1098/rsfs.2011.0047}, number=6, journal={Interface Focus}, publisher={The Royal Society}, author={Golightly, Andrew and Wilkinson, Darren J.}, year=2011, month=sep, pages={807–820} }
+@article{Golightly2011, title={Bayesian parameter inference for stochastic biochemical network models using particle {M}arkov chain {M}onte {C}arlo}, volume=1, ISSN={2042-8901}, url={http://dx.doi.org/10.1098/rsfs.2011.0047}, DOI={10.1098/rsfs.2011.0047}, number=6, journal={Interface Focus}, publisher={The Royal Society}, author={Golightly, Andrew and Wilkinson, Darren J.}, year=2011, month=sep, pages={807–820} }
 
 @software{JAX,
   author = {James Bradbury and Roy Frostig and Peter Hawkins and Matthew James Johnson and Chris Leary and Dougal Maclaurin and George Necula and Adam Paszke and Jake Vander{P}las and Skye Wanderman-{M}ilne and Qiao Zhang},
   title = {{JAX}: composable transformations of {P}ython+{N}um{P}y programs},
-  url = {http://github.com/google/jax},
+  url = {http://github.com/jax-ml/jax},
[…]
