+++ Makefile
@@ -1,6 +1,6 @@
 # Makefile
 
-VERSION=1.1.5
+VERSION=1.1.6
 
 FORCE:
 	make install
