### 2025-02-04 #7491 @darrenjw
OK - I think that is all done.
The new version is v1.1.5, and I've archived it on zenodo as: https://doi.org/10.5281/zenodo.14803573
