### 2025-02-04 #7491 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
